Crate cumulus_test_relay_validation_worker_provider

Provides the VALIDATION_WORKER for integration tests in Cumulus.

The validation worker is used by the relay chain to validate parachains. This worker is placed in an extra process to provide better security and to ensure that a worker can be killed etc.

!!This should only be used for tests!!

Re-exports

pub use polkadot_node_core_pvf;

Constants

VALIDATION_WORKER

The path to the validation worker.